backend_fixed/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.6.3
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.3
httpx==0.27.0
tenacity==8.2.3
python-dotenv==1.0.1
slowapi==0.1.9
structlog==24.1.0
PyYAML==6.0.1